fix(Analyser): better mimic the SuppressMessageAttribute
