+++ tests/Compiler/Analyser/Suppression.cs
@@ -16,7 +16,7 @@ public class SuppressionAttributeTests {
         var attributes = TestData.GetAttributes(astContent);
         Assert.Multiple(() => {
             foreach (var attribute in attributes) {
-                var result = SuppressAnalyserAttribute.FromAttributeAst(attribute).Unwrap();
+                var result = SuppressAnalyserAttributeExt.FromAttributeAst(attribute).Unwrap();
                 Assert.That(result, Is.EqualTo(Option<SuppressAnalyserAttribute>.None));
             }
         });
@@ -27,10 +27,10 @@ public class SuppressionAttributeTests {
         var attributes = TestData.GetAttributes(astContent);
         var i = 0;
         foreach (var attribute in attributes) {
-            var result = SuppressAnalyserAttribute.FromAttributeAst(attribute).Unwrap();
+            var result = SuppressAnalyserAttributeExt.FromAttributeAst(attribute).Unwrap();
             if (!result.IsSome(out var attr)) continue;
             Assert.Multiple(() => {
-                var suppression = attr.GetSupression();
+                var suppression = attr.GetSuppression();
 
                 Assert.That(suppression?.Type, Is.EqualTo(expected[i].Type));
                 Assert.That(suppression?.Justification, Is.EqualTo(expected[i].Justification));
@@ -121,10 +121,14 @@ public class SuppressionAttributeTests {
 
                 yield return new TestCaseData("""
                 function Test-Function {
-                    [SuppressAnalyser('UseOfUndefinedFunction', 'This-Function', 'This Justification')]
-                    [SuppressAnalyserAttribute('UseOfUndefinedFunction', 'That-Function', 'That Justification')]
-                    [Compiler.Analyser.SuppressAnalyser('UseOfUndefinedFunction', 'Other-Function', 'Other Justification')]
-                    [Compiler.Analyser.SuppressAnalyserAttribute('UseOfUndefinedFunction', 'Another-Function', 'Another Justification')]
+                    [SuppressAnalyser('UseOfUndefinedFunction', 'This-Function', Justification = 'This Justification')]
+                    [SuppressAnalyserAttribute('UseOfUndefinedFunction', 'That-Function', Justification = 'That Justification')]
+                    [Compiler.Analyser.SuppressAnalyser('UseOfUndefinedFunction', 'Other-Function', Justification = 'Other Justification')]
+                    [Compiler.Analyser.SuppressAnalyserAttribute('UseOfUndefinedFunction', 'Another-Function', Justification = 'Another Justification')]
+                    [SuppressAnalyser('UseOfUndefinedFunction', 'Other-Function')]
+                    [SuppressAnalyserAttribute('UseOfUndefinedFunction', 'Another-Function')]
+                    [Compiler.Analyser.SuppressAnalyser('UseOfUndefinedFunction', 'This-Function')]
+                    [Compiler.Analyser.SuppressAnalyserAttribute('UseOfUndefinedFunction', 'That-Function')]
                     param()
                 }
                 """,
@@ -132,7 +136,11 @@ public class SuppressionAttributeTests {
                     new(typeof(UseOfUndefinedFunction), "This-Function", "This Justification"),
                     new(typeof(UseOfUndefinedFunction), "That-Function", "That Justification"),
                     new(typeof(UseOfUndefinedFunction), "Other-Function", "Other Justification"),
-                    new(typeof(UseOfUndefinedFunction), "Another-Function", "Another Justification")
+                    new(typeof(UseOfUndefinedFunction), "Another-Function", "Another Justification"),
+                    new(typeof(UseOfUndefinedFunction), "This-Function", null),
+                    new(typeof(UseOfUndefinedFunction), "That-Function", null),
+                    new(typeof(UseOfUndefinedFunction), "Other-Function", null),
+                    new(typeof(UseOfUndefinedFunction), "Another-Function", null)
                 });
             }
         }
